% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accident_2013.R
\docType{data}
\name{accident_2013.csv.bz2}
\alias{accident_2013.csv.bz2}
\title{2013 Yearly accidents reports dataset}
\format{
A \code{tibble} with 30202
observations and 50 columns
\describe{   # Documenting only very few columns
\item{YEAR}{Year when accident occurred}
\item{MONTH}{Month when accident occurred}
\item{DAY}{Day when accident occurred}
\item{CITY}{City where accident occurred}
\item{LONGITUD}{Longitud of location where accident occurred}
\item{LATITUDE}{Latitude of location where accident occurred}
}
}
\source{
From US NHTSA website
\url{https://www.nhtsa.gov/Data/Fatality-Analysis-Reporting-System-(FARS)}
}
\usage{
accident_2013.csv.bz2
}
\description{
2013 Yearly accidents reports dataset
}
\examples{
\dontrun{
accident_2013.csv.bz2

"accident_2013.csv.bz2"
}
}
\keyword{datasets}
