% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_map_state.R
\name{fars_map_state}
\alias{fars_map_state}
\title{draw a geographical map and add points to it}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num, }{\code{numeric}, representing a state number}

\item{year, }{\code{numeric}, a number representing a year}
}
\value{
This function returns \verb{a geographical map containing points} where
accidents occurred
}
\description{
Both arguments \code{state.num} and \code{year} are coerced to integers.
\code{year}is used to create a file name. The \code{state.num}is used to filter
that file at the variable \code{STATE}. Using the state associated with
\code{state.num} as database, a map is drawn. Using LONGITUD and LATITUDE as
coordinates vectors, \code{fars_map_state} adds points, locations where
accidents occurred to the map.

\code{fars_map_state} returns error if any of the following conditions is true:
\itemize{
\item any of its arguments cannot be coerced to integer
\item the file doesn't exist for the file name created with \code{year}
\item \code{state.num} is not a value in the column \code{STATE} of the file associated
with \code{year}.
}
}
\examples{
\dontrun{
fars_map_state(1, 2013)
fars_map_state(56, 2015)
}

}
