% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read.R
\name{fars_read}
\alias{fars_read}
\title{Reads file into R as a tibble}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A string, name of a file to read into R}
}
\value{
The \code{fars_read} function returns the file named file name as
\verb{a tibble}.
}
\description{
\code{fars_read} takes \verb{a string} argument used as
file name.If a file with that name exists, it reads that file into R
without showing progress bars, ignoring all simple diagnostic messages
and coerces it to \verb{a tibble}. As side effect, \code{fars_read} prints
out the file as \code{tibble}.

If \code{filename} is not a name of a file or is not a string,\code{fars_read}
returns an error.
}
\examples{
\dontrun{
fars_read(accident_2015.csv.bz2)
fars_read(make_filename(2013))
fars_read(make_filename(2014))
}

}
