% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read_years.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Create 2 columns tibble (MONTH, year) for each element of a
list or numeric vector of years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{list or numeric vector containing years as integers}
}
\value{
This function returns \verb{a list} of 2 columns tibbles. Numbers of rows
are the same as in the file of each element of the argument years
}
\description{
The \code{fars_read_years} function takes a list or numeric vector
argument \code{years}and uses the function \code{make_filename}to create a \verb{file name for each element }year\verb{ in}years\verb{. Then returns a two columns tibble: MONTH, year for each }year` by adding a new column (year) to the file of that
year and selecting the columns MONTH, year.

\code{fars_read_years} produces an error if any element \code{year} in the argument
\code{years} cannot be coerced to integer or if there is no file for the \verb{file name created for that}year`.
}
\examples{
\dontrun{
fars_read_years(c(2013, 2014, 2015))
fars_read_years(c(2013, 2015))
}

}
