% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_summarize_years.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{number of accidents in each month for a list or numeric vector of years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A list or numeric vector containing years as integers}
}
\value{
\verb{a tibble}, summary of numbers of accidents in each month of
each year in the argument \code{years} supplied.
12 rows and (1 + number of elements in years) columns.
}
\description{
\code{fars_summarize_years} function takes a list or numeric vector
argument \code{years}. It creates a list of 2 columns tibbles: MONTH, year
for each element \code{year}in \code{years}. It Row-binds, regroups and summarizes
these 2 columns tibbles to return summary of numbers of accidents in each
month of each element of \code{years}.

\code{fars_summarize_years} will generate error if any element of \code{years} cannot
be coerced to integer or if there is no file for the file name created with
any element of \code{years}.
}
\examples{
\dontrun{
fars_summarize_years(c(2013, 2014, 2015))
fars_summarize_years(c(2013, 2015))
}

}
