% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_filename.R
\name{make_filename}
\alias{make_filename}
\title{Create a file name using an integer}
\usage{
make_filename(year)
}
\arguments{
\item{year}{numeric, integer referencing a year}
}
\value{
The function \code{make_filename} returns \verb{a string}.
}
\description{
\code{make_filename} takes a numeric argument \code{year}that it coerces
to integer and supplies it along with\code{fmt = "accident_\%d.csv,bz2"} to the
\code{sprintf} function to create the file name by replacing \verb{\%d} by \code{year}.

\code{make_filename} will generate an error if its argument cannot be coerced
to integer.
}
\examples{
\dontrun{
 make_filename(2015)
 make_filename(2013)
 make_filename(001)
 make_filename(20.75)
}

}
